Conflicts:
	assembler/src/debruijn/path_extend/paired_library.hpp
	assembler/src/debruijn/repeat_resolving_routine.hpp

+++ assembler/src/spades_pipeline/spades_logic.py
@@ -15,22 +15,23 @@ import support
 import process_cfg
 from process_cfg import bool_to_str
 from process_cfg import load_config_from_file
+import options_storage
 
-def prepare_config_spades(filename, cfg, log, prev_K, K, last_one):
+def prepare_config_spades(filename, cfg, log, use_additional_contigs, K, last_one):
     subst_dict = dict()
 
     subst_dict["K"] = str(K)
     subst_dict["run_mode"] = "false"
-    subst_dict["dataset"] = cfg.dataset
-    subst_dict["output_base"] = cfg.output_dir
-    subst_dict["additional_contigs"] = cfg.additional_contigs
+    subst_dict["dataset"] = process_cfg.process_spaces(cfg.dataset)
+    subst_dict["output_base"] = process_cfg.process_spaces(cfg.output_dir)
+    subst_dict["additional_contigs"] = process_cfg.process_spaces(cfg.additional_contigs)
     subst_dict["entry_point"] = "construction"
     subst_dict["developer_mode"] = bool_to_str(cfg.developer_mode)
     subst_dict["gap_closer_enable"] = bool_to_str(last_one)
     subst_dict["paired_mode"] = bool_to_str(last_one and cfg.paired_mode)
     subst_dict["long_single_mode"] = bool_to_str(last_one and cfg.long_single_mode)
     subst_dict["topology_simplif_enabled"] = bool_to_str(last_one)
-    subst_dict["use_additional_contigs"] = bool_to_str(prev_K)
+    subst_dict["use_additional_contigs"] = bool_to_str(use_additional_contigs)
     subst_dict["max_threads"] = cfg.max_threads
     subst_dict["max_memory"] = cfg.max_memory
     subst_dict["correct_mismatches"] = bool_to_str(last_one)
@@ -38,12 +39,15 @@ def prepare_config_spades(filename, cfg, log, prev_K, K, last_one):
         subst_dict["resolving_mode"] = cfg.resolving_mode
     if "careful" in cfg.__dict__:
         subst_dict["mismatch_careful"] = bool_to_str(cfg.careful)
+    if "pacbio_mode" in cfg.__dict__:
+        subst_dict["pacbio_test_on"] = bool_to_str(cfg.pacbio_mode)
+        subst_dict["pacbio_reads"] = process_cfg.process_spaces(cfg.pacbio_reads)
 
     process_cfg.substitute_params(filename, subst_dict, log)
 
 
 def get_read_length(output_dir, K):
-    estimated_params = load_config_from_file(os.path.join(output_dir, "K%d" % (K), "_est_params.info"))
+    estimated_params = load_config_from_file(os.path.join(output_dir, "K%d" % K, "_est_params.info"))
     lib_count = int(estimated_params.__dict__["lib_count"])
     max_read_length = 0
     for i in range(lib_count):
@@ -52,9 +56,28 @@ def get_read_length(output_dir, K):
     return max_read_length
 
 
+def update_k_mers_in_special_cases(cur_k_mers, RL, log):
+    if not options_storage.k_mers and not options_storage.single_cell: # kmers were set by default and not SC
+        if RL >= 250:
+            support.warning("Default k-mer sizes were set to %s because estimated "
+                            "read length (%d) is equal or great than 250" % (str(options_storage.k_mers_250), RL), log)
+            return options_storage.k_mers_250
+        if RL >= 150:
+            support.warning("Default k-mer sizes were set to %s because estimated "
+                            "read length (%d) is equal or great than 150" % (str(options_storage.k_mers_150), RL), log)
+            return options_storage.k_mers_150
+    return cur_k_mers
+
 
 def run_iteration(configs_dir, execution_home, cfg, log, K, use_additional_contigs, last_one):
-    data_dir = os.path.join(cfg.output_dir, "K%d" % (K))
+    data_dir = os.path.join(cfg.output_dir, "K%d" % K)
+    if options_storage.continue_mode:
+        if os.path.isfile(os.path.join(data_dir, "final_contigs.fasta")):
+            log.info("\n== Skipping assembler: " + ("K%d" % K) + " (already processed)")
+            return
+        else:
+            options_storage.continue_mode = False # continue from here
+
     if os.path.exists(data_dir):
         shutil.rmtree(data_dir)
     os.makedirs(data_dir)
@@ -74,10 +97,9 @@ def run_iteration(configs_dir, execution_home, cfg, log, K, use_additional_conti
                     os.rename(cfg_file, cfg_file.split('.template')[0])
 
     prepare_config_spades(cfg_file_name, cfg, log, use_additional_contigs, K, last_one)
-    prev_K = K
 
-    command = os.path.join(execution_home, "spades") + " " +\
-               os.path.abspath(cfg_file_name)
+    command = [os.path.join(execution_home, "spades"),
+               os.path.abspath(cfg_file_name)]
 
     if os.path.isdir(bin_reads_dir):
         if glob.glob(os.path.join(bin_reads_dir, "*_cor*")):
@@ -86,7 +108,7 @@ def run_iteration(configs_dir, execution_home, cfg, log, K, use_additional_conti
                 new_bin_filename = cor_filename[:cor_index] + cor_filename[cor_index + 4:]
                 shutil.move(cor_filename, new_bin_filename)
 
-    log.info("\n== Running assembler: " + ("K%d" % (K)) + "\n")
+    log.info("\n== Running assembler: " + ("K%d" % K) + "\n")
     support.sys_call(command, log)
 
 
@@ -96,7 +118,7 @@ def run_spades(configs_dir, execution_home, cfg, log):
     cfg.iterative_K = sorted(cfg.iterative_K)
 
     bin_reads_dir = os.path.join(cfg.output_dir, ".bin_reads")
-    if os.path.isdir(bin_reads_dir):
+    if os.path.isdir(bin_reads_dir) and not options_storage.continue_mode:
         shutil.rmtree(bin_reads_dir)
 
     if len(cfg.iterative_K) == 1:
@@ -105,9 +127,12 @@ def run_spades(configs_dir, execution_home, cfg, log):
     else:
         run_iteration(configs_dir, execution_home, cfg, log, cfg.iterative_K[0], False, False)
         RL = get_read_length(cfg.output_dir, cfg.iterative_K[0])
-        if (cfg.iterative_K[1] > RL):
+        cfg.iterative_K = update_k_mers_in_special_cases(cfg.iterative_K, RL, log)
+        if cfg.iterative_K[1] + 1 > RL:
             if cfg.paired_mode:
-                log.info("Second value of iterative K exceeded estimated read length. Rerunning in paired mode for the first value of K")
+                support.warning("Second value of iterative K (%d) exceeded estimated read length (%d). "
+                                "Rerunning in paired mode for the first value of K (%d)" %
+                                (cfg.iterative_K[1], RL, cfg.iterative_K[0]), log)
                 run_iteration(configs_dir, execution_home, cfg, log, cfg.iterative_K[0], False, True)
                 K = cfg.iterative_K[0]
         else:
@@ -116,22 +141,26 @@ def run_spades(configs_dir, execution_home, cfg, log):
             count = 0
             for K in rest_of_iterative_K:
                 count += 1
-                last_one = count == len(cfg.iterative_K) or rest_of_iterative_K[count] > RL
+                last_one = count == len(cfg.iterative_K) or (rest_of_iterative_K[count] + 1 > RL)
                 run_iteration(configs_dir, execution_home, cfg, log, K, True, last_one)
                 if last_one:
                     break
             if count < len(cfg.iterative_K):
-                log.info("Iterations stopped. Value of K exceeded estimated read length")
+                support.warning("Iterations stopped. Value of K (%d) exceeded estimated read length (%d)" %
+                                (cfg.iterative_K[count], RL), log)
 
-    latest = os.path.join(cfg.output_dir, "K%d" % (K))
+    latest = os.path.join(cfg.output_dir, "K%d" % K)
 
     if os.path.isfile(os.path.join(latest, "before_rr.fasta")):
-        shutil.copyfile(os.path.join(latest, "before_rr.fasta"), os.path.join(os.path.dirname(cfg.result_contigs), "before_rr.fasta"))
-    if os.path.isfile(os.path.join(latest, "final_contigs.fasta")):    
-        shutil.copyfile(os.path.join(latest, "final_contigs.fasta"), cfg.result_contigs)
+        if not os.path.isfile(os.path.join(os.path.dirname(cfg.result_contigs), "before_rr.fasta")) or not options_storage.continue_mode:
+            shutil.copyfile(os.path.join(latest, "before_rr.fasta"), os.path.join(os.path.dirname(cfg.result_contigs), "before_rr.fasta"))
+    if os.path.isfile(os.path.join(latest, "final_contigs.fasta")):
+        if not os.path.isfile(cfg.result_contigs) or not options_storage.continue_mode:
+            shutil.copyfile(os.path.join(latest, "final_contigs.fasta"), cfg.result_contigs)
     if cfg.paired_mode:
         if os.path.isfile(os.path.join(latest, "scaffolds.fasta")):
-            shutil.copyfile(os.path.join(latest, "scaffolds.fasta"), cfg.result_scaffolds)
+            if not os.path.isfile(cfg.result_scaffolds) or not options_storage.continue_mode:
+                shutil.copyfile(os.path.join(latest, "scaffolds.fasta"), cfg.result_scaffolds)
 
     if cfg.developer_mode:
         # before repeat resolver contigs
@@ -147,7 +176,5 @@ def run_spades(configs_dir, execution_home, cfg, log):
 
     if os.path.isdir(bin_reads_dir):
         shutil.rmtree(bin_reads_dir)
-    if not cfg.paired_mode:
-        return cfg.result_contigs, "", latest
 
-    return cfg.result_contigs, cfg.result_scaffolds, latest
+    return latest
